+++ src/main/kotlin/models/Category.kt
@@ -3,4 +3,4 @@ package org.example.models
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Category(val id: Int? = null, val name: String? = null)
+data class Category(val id: Long? = null, val name: String? = null)
